feat(components): Enhance ActiveOffers with BusinessCard component
- Extract business card rendering logic into separate BusinessCard component
- Improve code modularity and readability in ActiveOffers
- Add new BusinessCard component with dynamic rendering of business details
- Create BusinessCard.test.tsx for component testing
- Refactor location and payment method display logic
- Maintain existing UI/UX patterns and interaction behaviors
